+++ src/tailrecursion/hoplon/compiler/compiler.clj
@@ -8,15 +8,15 @@
 
 (ns tailrecursion.hoplon.compiler.compiler
   (:require
-    [clojure.java.io                        :refer  [file make-parents]]
+    [clojure.java.io                        :refer  [file resource make-parents]]
     [clojure.pprint                         :as     pprint]
     [clojure.string                         :as     string]
     [cljs.compiler                          :as     cljsc]
     [tailrecursion.javelin                  :refer  [make-require make-require-macros]]
     [tailrecursion.hoplon.compiler.tagsoup  :as     ts]))
 
-(def ^:dynamic *current-file* nil)
-(def ^:dynamic *printer*      prn)
+(def ^:dynamic *printer* prn)
+(def file->page (atom {}))
 
 (def html-tags
   '[a abbr acronym address applet area article
@@ -38,17 +38,21 @@
     (->> [name] (concat (repeat (count dirs) "../")) (apply str))))
 
 (defn munge-path [path]
-  (-> (str "__" path) (string/replace "_" "__") (string/replace "/" "_")))
+  (-> (str "__" path)
+    (string/replace "_" "__")
+    (string/replace "/" "_")
+    (string/replace ":" "__COLON__")))
 
-(defn as-forms-string [s type]
-  (cond (= type "html") (-> s ts/parse-string ts/tagsoup->hoplon)
-        (= type "cljs") (read-string (str "(" s ")"))))
-
-(defn as-forms-path [path]
-  (as-forms-string (slurp path) (last (re-find #"[^.]+\.([^.]+)\.hl$" path))))
+(defn as-forms [s]
+  (case (first (string/trim s))
+    \< (-> s ts/parse-string ts/tagsoup->hoplon)
+    (try
+    (read-string (str "(" s ")"))
+      (catch Throwable e (println s))
+      )))
 
 (defn output-path     [forms] (-> forms first second str))
-(defn output-path-for [path]  (-> path as-forms-path output-path))
+(defn output-path-for [path]  (-> path slurp as-forms output-path))
 
 (defn make-nsdecl
   [[_ ns-sym & forms]]
@@ -63,38 +67,57 @@
         others  (->> forms (filter list?) (filter other?))]
     `(~'ns ~ns-sym ~@others ~reqs ~macros)))
 
+(defn forms-str [forms]
+  (string/join "\n" (map #(with-out-str (*printer* %)) forms)))
+
+(defn compile-lib [[[ns* & _ :as nsdecl] & tlfs]]
+  (when (= 'ns ns*) (forms-str (cons (make-nsdecl nsdecl) tlfs))))
+
 (defn compile-forms [forms js-file]
-  (let [[[nsdecl & setup] html] ((juxt butlast last) forms)
-        outpath   (output-path forms)
-        js-uri    (up-parents outpath (.getName js-file))
-        nsdecl    (let [n (make-nsdecl nsdecl)]
-                    (concat (take 1 n) [(munge (second n))] (drop 2 n))) 
-        nsname    (cljsc/munge (second nsdecl)) 
-        s-nodep   (list 'script {:type "text/javascript"} "var CLOSURE_NO_DEPS = true;")
-        s-main    (list 'script {:type "text/javascript" :src js-uri})
-        s-init    (list 'script {:type "text/javascript"} (str nsname ".hoploninit();"))
-        s-html    (list 'html
-                        (list 'head (list 'meta {:charset "utf-8"}))
-                        (list 'body s-nodep s-main s-init))
-        htmlstr   (ts/pp-forms "html" s-html)
-        cljs      (concat
-                    (list nsdecl)
-                    (list
-                      (list 'defn (symbol "^:export") 'hoploninit []
-                            (cons 'do setup) 
-                            (list (symbol "tailrecursion.hoplon/init") html))))
-        cljsstr (string/join "\n" (map #(with-out-str (*printer* %)) cljs))]
-    {:html htmlstr :cljs cljsstr :file outpath}))
+  (let [[nsdecl & tlfs] forms]
+    (if (= 'ns (first nsdecl))
+      {:cljs (forms-str (cons (make-nsdecl nsdecl) tlfs))}
+      (let [[[_ & setup] html] ((juxt butlast last) forms)
+            outpath   (output-path forms)
+            js-uri    (up-parents outpath (.getName js-file))
+            mkns      #(symbol (str "tailrecursion.hoplon.app-pages." (munge %)))
+            nsdecl    (let [[h n & t] (make-nsdecl nsdecl)] (list* h (mkns n) t))
+            nsname    (cljsc/munge (second nsdecl)) 
+            s-nodep   (list 'script {:type "text/javascript"} "var CLOSURE_NO_DEPS = true;")
+            s-main    (list 'script {:type "text/javascript" :src js-uri})
+            s-init    (list 'script {:type "text/javascript"} (str nsname ".hoploninit();"))
+            s-html    (list 'html
+                            (list 'head (list 'meta {:charset "utf-8"}))
+                            (list 'body s-nodep s-main s-init))
+            htmlstr   (ts/pp-forms "html" s-html)
+            cljs      (list nsdecl
+                            (list 'defn (symbol "^:export") 'hoploninit []
+                                  (cons 'do setup) 
+                                  (list (symbol "tailrecursion.hoplon/init") html)))
+            cljsstr   (forms-str cljs)]
+        {:html htmlstr :cljs cljsstr :file outpath}))))
 
 (defn pp [form] (pprint/write form :dispatch pprint/code-dispatch))
 
-(defn compile-file
-  [f js-file cljsdir htmldir & {:keys [opts]}]
-  (when-let [forms (as-forms-path (.getPath f))]
-    (binding [*current-file*  f
-              *printer*       (if (:pretty-print opts) pp prn)]
+(defn delete-path [p]
+  (doseq [[k v] (get @file->page p)]
+    (when (and v (.exists v)) (.delete v)))
+  (swap! file->page dissoc p))
+
+(defn compile-string
+  [forms-str path js-file cljsdir htmldir & {:keys [opts]}]
+  (when-let [forms (as-forms forms-str)]
+    (binding [*printer* (if (:pretty-print opts) pp prn)]
       (let [compiled (compile-forms forms js-file)
-            cljs-out (file cljsdir (str (munge-path (.getPath f)) ".cljs"))
-            html-out (file htmldir (:file compiled))
-            write    #(spit (doto %1 make-parents) %2)]
-        (doall (map write [cljs-out html-out] ((juxt :cljs :html) compiled)))))))
+            cljs-out (file cljsdir (str (munge-path path) ".cljs"))
+            html-out (when-let [f (:file compiled)] (file htmldir f))
+            write    #(when (and %1 %2) (spit (doto %1 make-parents) %2))]
+        (delete-path path)
+        (swap! file->page assoc path {:cljs cljs-out :html html-out})
+        (write cljs-out (:cljs compiled))
+        (write html-out (:html compiled))))))
+
+(defn compile-file [f & args]
+  (if (and (instance? java.io.File f) (not (.exists f)))
+    (delete-path (.getPath f))
+    (apply compile-string (slurp f) (.getPath f) args)))